Add two testng suite configuration to support smoke and regression tests
