--- src/main/resources/TestNG.xml
@@ -1,4 +1,0 @@
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="package">
-
-</suite>
